SkillFinderWeb/SkillFinderWeb/requirements.txt
azure-storage==0.35.1
beautifulsoup4==4.6.0
bs4==0.0.1
certifi>=2017.4.17
chardet>=3.0.4
click>=6.7
Flask<1
Flask-OAuthlib>=0.9.4
idna>=2.5
itsdangerous>=0.24
Jinja2>=2.9.6
MarkupSafe>=1.0
oauthlib>=2.0.2
python-arango==3.9.0
requests>=2.18.1
requests-oauthlib>=0.8.0
six==1.10.0
urllib3>=1.21.1
Werkzeug>=0.12.2